+++ simplegen/src/test/kotlin/com/ancientlightstudios/simplegen/TemplateEngineTests.kt
@@ -39,7 +39,7 @@ class TemplateEngineTests : BehaviorSpec({
 
     Given("i have a template with a jsonpath filter") {
 
-        val data = YamlReader.readToMap("plain text", "nestedprop:\n  prop2: value2\n".byteInputStream())
+        val data = YamlParser().parse("nestedprop:\n  prop2: value2\n".byteInputStream(), "plain text")
         val context = mapOf<String, Any>("json" to data)
         val template = "{{ json | jsonpath('nestedprop.prop2') }}"
         val engine = TemplateEngine(getResourcesRootFileResolver())